pyproject.toml
[project]
name = "switch-operator"
version = "0.1.0"
description = "To be used with ESP32C3 based servo controllers"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi>=0.115.13",
    "jinja2>=3.1.6",
    "paho-mqtt>=2.1.0",
    "pydantic>=2.11.7",
    "python-multipart>=0.0.20",
    "uvicorn>=0.34.3",
]

[dependency-groups]
dev = [
    "pytest>=8.4.1",
    "ruff>=0.12.0",
]
